Fix pylint comments
