+++ art/attacks/projected_gradient_descent.py
@@ -79,7 +79,7 @@ class ProjectedGradientDescent(FastGradientMethod):
 
         self._project = True
 
-    def generate(self, x, y=None):
+    def generate(self, x, y=None, **kwargs):
         """
         Generate adversarial samples and return them in an array.
