+++ art/attacks/spatial_transformation.py
@@ -73,7 +73,7 @@ class SpatialTransformation(Attack):
         self.attack_trans_y = None
         self.attack_rot = None
 
-    def generate(self, x, y=None):
+    def generate(self, x, y=None, **kwargs):
         """
         Generate adversarial samples and return them in an array.
