+++ art/attacks/virtual_adversarial.py
@@ -62,7 +62,7 @@ class VirtualAdversarialMethod(Attack):
                   'batch_size': batch_size}
         self.set_params(**kwargs)
 
-    def generate(self, x, y=None):
+    def generate(self, x, y=None, **kwargs):
         """
         Generate adversarial samples and return them in an array.
