+++ tests/attacks/test_boundary.py
@@ -26,8 +26,9 @@ import tensorflow as tf
 
 from art.attacks import BoundaryAttack
 from art.classifiers import KerasClassifier
-from art.utils import load_dataset, random_targets, master_seed, get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from art.utils import load_dataset, random_targets, master_seed
+from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger('testLogger')
