+++ tests/attacks/test_elastic_net.py
@@ -26,8 +26,9 @@ import tensorflow as tf
 
 from art.attacks import ElasticNet
 from art.classifiers import KerasClassifier
-from art.utils import load_dataset, random_targets, master_seed, get_classifier_tf, get_classifier_kr
-from art.utils import get_classifier_pt, get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from art.utils import load_dataset, random_targets, master_seed
+from art.utils_test import get_classifier_tf, get_classifier_kr
+from art.utils_test import get_classifier_pt, get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger('testLogger')
