+++ tests/attacks/test_saliency_map.py
@@ -27,8 +27,8 @@ import tensorflow as tf
 from art.attacks.saliency_map import SaliencyMapMethod
 from art.classifiers import KerasClassifier
 from art.utils import load_dataset, get_labels_np_array, to_categorical, master_seed
-from art.utils import get_classifier_tf, get_classifier_kr, get_classifier_pt
-from art.utils import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
+from art.utils_test import get_classifier_tf, get_classifier_kr, get_classifier_pt
+from art.utils_test import get_iris_classifier_tf, get_iris_classifier_kr, get_iris_classifier_pt
 
 logger = logging.getLogger('testLogger')
